Docs: Created an extension of MapViewController and added 3 mapView methods
The first method will detect if the region has been changed, and if so, will send those coordinates and deltas to the mapSet document. The viewForAnnotation method will animate the dropping of the pin, and the did select will check if the user selected edit or not and handles the event accordingly after selecting a pin annotation.