fix: Adjusted the location of the Photo entity records
At the .end case of my switch statement for the LongGestureRecognizer, I have called the Network request to the Flickr Client and added the 'url_m' to the Photo attribute MURL. Adjusted the GetPhotos() in the PhotoVC. Does the same thing as the end case of the switch statement in MapVC. Appending the photos to the Photo NSSet attribute of the PinAnnotation record. Not appending!